cli/fs: just check if dir exists for submodule
The previous check would only return false if the dir didn't exist. If
it did, it would be inside the repository we initialized earlier, so it
would indeed be inside a Git worktree. And checking for a Git worktree
on a directory that doesn't exist seems to error on Windows.