fix inconsistency between manifest and compiled package config
During compilation, dependencies' configs must be updated, not only in
the CLI side when generating the manifest or reading the root config.

Otherwise, we'd get a warning saying that the dependency is importing a
module from something that it doesn't depend on, which it can only
suppose to be a transitive dependency.