Improved dependency patching (#44)
* patching attempt 2 - beginnings

* rename replacements to glistix_replacements for clarity

help when merging stuff in the future

* patch manifest requirements

needed for the locked() function

* fix inconsistency between manifest and compiled package config

During compilation, dependencies' configs must be updated, not only in
the CLI side when generating the manifest or reading the root config.

Otherwise, we'd get a warning saying that the dependency is importing a
module from something that it doesn't depend on, which it can only
suppose to be a transitive dependency.

* default config read to patching

* rename replacements to patches

* test glistix patches in dependency resolving

* make patch rename optional

* glistix patch: allow not specifying name, flatten source

* send unpatched hex dependencies when publishing

Ideally docs would be unpatched as well, but then we can't build.

* add patches to manifest

needed to know if they changed

* always unlock packages affected by new patches

For example, if we patch gleam_stdlib and rename it to glistix_stdlib,
now any packages depending on gleam_stdlib OR glistix_stdlib, as well as
both packages themselves (if they appear in the manifest), should be
considered stale and removed from the manifest.

* consider removed patches when unlocking

* use patch in 'glistix new'

Back out "cli/new: automatically clone stdlib,gleeunit"

This backs out commit c7b59c23262fa2477bca3e8a4f6d8c6fc2ba7ce6.

* clippy and update tests

* fix windows tests

* don't make patched hex dependency a root dependency unless necessary

* add glistix_project_nix test

* add error hints suggesting patches

* improve link display in error message

* revert some clippy changes

* prompt version resolving when patches differ

* further changes to default gleam.toml

- remove hex-patch suggestion

* add test/glistix_project_nix to CI

* improve renamedpkg example in gleam.toml

* add 'maybe forgot patch hint' to unsupported nix target feature

* check for 'build/packages' for maybe forgot patch error hint

* add a comment to manifest requirement renaming

* regen manifest after 'glistix add' if patched dep was added

* remove unused patching functions

* add todo for better unlocking behavior in the future

* fix glistix new giving wrong version override suggestion

* add config patch tests

* fix: also patch dev deps

* fix: check for patch conflict, deterministic patch order

* small spacing changes in config.rs

* fix patching with local deps using relative paths

* add local patching test

* add glistix_hex_patching test

* fix unnecessary clone and add todo

* discourage local-overrides setting